pyproject.toml
[tool.poetry]
name = "mytest"
version = "0.1.0"
description = ""
authors = ["MIT"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.11"}
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
aiosqlite = "^0.21.0"
sqlalchemy = "^2.0.38"
uvicorn = "^0.34.0"
alembic = "^1.15.1"
pyjwt = "2.2.0"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
asyncpg = "^0.30.0"
aiobotocore = "^2.22.0"
python-dotenv = "^1.1.0"
cryptography = "^45.0.3"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
